+++ Libplanet.Tests/Store/StoreTest.cs
@@ -932,6 +932,10 @@ namespace Libplanet.Tests.Store
             Assert.Equal(Fx.Block2.Hash, store.IndexBlockHash(chainB, 2));
             Assert.Equal(Fx.Block2.Hash, store.IndexBlockHash(chainC, 2));
             Assert.Equal(Fx.Block3.Hash, store.IndexBlockHash(chainC, 3));
+
+            Assert.Throws<ChainIdNotFoundException>(
+                () => store.ForkBlockIndexes(Guid.NewGuid(), Guid.NewGuid(), Fx.Block2.Hash)
+            );
         }
 
         [SkippableFact]